pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "schema-translator"
version = "0.1.0"
description = "LLM-powered semantic schema mapping for heterogeneous tenant data"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [
    {name = "Schema Translator Team"},
]
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "pandas>=2.1.0",
    "numpy>=1.24.0",
    "pyyaml>=6.0",
    "openai>=1.3.0",
    "python-multipart>=0.0.6",
    "typer>=0.9.0",
    "rich>=13.0.0",
    "python-dateutil>=2.8.0",
    "fuzzywuzzy>=0.18.0",
    "python-levenshtein>=0.23.0",
    "pyarrow>=14.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "mypy>=1.7.0",
]

[project.scripts]
schema-translator = "src.app.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["src/app"]

[tool.black]
line-length = 88
target-version = ['py39']

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

